consolidate finance dashboards and rename spotify handler
